pyproject.toml
[tool.poetry]
name = "prompt-share-backend"
version = "0.1.0"
description = "Your new great project"
authors = ["Afaneor"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
django-split-settings = "^1.1.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21"}
aerich = "0.7.1"
httpx = "^0.27"
pydantic-settings = "^2.4.0"
sentry-sdk = "^2.16.0"
fastapi-pagination = "^0.12.31"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.2"
requests = "^2.28.1"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"

[tool.aerich]
tortoise_orm = "config.tortoise_settings"
location = "app/db/migrations"
src_folder = "./app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
